[Revalidate] Create a new scope for each revalidation
