+++ src/NuGet.Services.Revalidate/NuGet.Services.Revalidate.csproj
@@ -58,20 +58,22 @@
     <Compile Include="Services\HealthService.cs" />
     <Compile Include="Services\IGalleryService.cs" />
     <Compile Include="Services\IHealthService.cs" />
-    <Compile Include="Services\IPackageRevalidationInserter.cs" />
+    <Compile Include="Initialization\IPackageRevalidationInserter.cs" />
     <Compile Include="Services\IRevalidationQueue.cs" />
     <Compile Include="Services\IRevalidationJobStateService.cs" />
     <Compile Include="Services\IPackageRevalidationStateService.cs" />
     <Compile Include="Services\IRevalidationService.cs" />
+    <Compile Include="Services\IRevalidationStarter.cs" />
     <Compile Include="Services\IRevalidationThrottler.cs" />
     <Compile Include="Services\ISingletonService.cs" />
-    <Compile Include="Services\PackageRevalidationInserter.cs" />
+    <Compile Include="Initialization\PackageRevalidationInserter.cs" />
     <Compile Include="Services\RevalidationOperation.cs" />
     <Compile Include="Services\RevalidationQueue.cs" />
     <Compile Include="Services\RevalidationResult.cs" />
     <Compile Include="Services\RevalidationService.cs" />
     <Compile Include="Services\RevalidationJobStateService.cs" />
     <Compile Include="Services\PackageRevalidationStateService.cs" />
+    <Compile Include="Services\RevalidationStarter.cs" />
     <Compile Include="Services\RevalidationThrottler.cs" />
     <Compile Include="Services\SingletonService.cs" />
     <Compile Include="Services\TelemetryService.cs" />
